Finished ecoli assembly stats
I did not include the alignments and misassembled contigs that QUAST
creates for the sake of saving space, but you'll find the rest of the
relevant data on the assembly here.

I suggest looking at the html report (the pdf is a bit too squished).
Also definitely check out the svg.