Error Finder
This is the skeleton script that outlines some of the things we need to
be working on as the assembly team.

This script accepts this arguments:
--sam FILE (sam file)
--coords FILE (coord file)
--mismatches FILE (snps file)

It's a good idea to use the sample files I've generated from my ecoli
dataset to test this out ( on Pinky, go to:
/data/assembly_team/test_data )

So far, this script can report reads that are in the wrong place, along
with whether or not they are near mismatches in the contigs (which we
assume were caused because of the wrong reads being used).

Look at main() to see an example of me playing around with some of the
functions..